change partition2Row to uint32_t

+++ cpp/core/shuffle/FallbackRangePartitioner.cc
@@ -23,7 +23,7 @@ namespace gluten {
 arrow::Status gluten::FallbackRangePartitioner::compute(
     const int32_t* pidArr,
     const int64_t numRows,
-    std::vector<uint16_t>& row2Partition,
+    std::vector<uint32_t>& row2Partition,
     std::vector<uint32_t>& partition2RowCount) {
   row2Partition.resize(numRows);
   std::fill(std::begin(partition2RowCount), std::end(partition2RowCount), 0);
