+++ cpp/core/shuffle/HashPartitioner.h
@@ -28,7 +28,7 @@ class HashPartitioner final : public Partitioner {
   arrow::Status compute(
       const int32_t* pidArr,
       const int64_t numRows,
-      std::vector<uint16_t>& row2partition,
+      std::vector<uint32_t>& row2partition,
       std::vector<uint32_t>& partition2RowCount) override;
 };
